display one chart at a time
