+++ src/components/SpeechDistribution/Sapogov.js
@@ -99,7 +99,17 @@ class Sapogov extends Component {
       options.scales.yAxes[0].ticks.stepSize = 1;
     }
 
-    return <Line data={data} options={options}/>;
+    return (
+      <>
+        <Line data={data} options={options}/>
+        <p>
+          {'Cf. '}
+          <a href="https://www.zotero.org/groups/940512/dlina/items/itemKey/BU7ZB3LY">
+            Sapogov 1974
+          </a>
+        </p>
+      </>
+    );
   }
 }
